+++ src/index.ts
@@ -4,9 +4,10 @@ export { TextDocument } from "./TextDocument";
 export { Image } from "./draw/Image";
 
 // style
+export { Color } from "./style/Color";
 export { HorizontalAlignment } from "./style/HorizontalAlignment";
-export { ITextStyle } from "./style/ITextStyle";
-export { Style } from "./style/Style";
+export { IParagraphStyle } from "./style/IParagraphStyle";
+export { ParagraphStyle } from "./style/ParagraphStyle";
 export { TabStop } from "./style/TabStop";
 export { TabStopType } from "./style/TabStopType";
 export { Typeface } from "./style/Typeface";