+++ src/style/StyleHelper.ts
@@ -1,37 +1,20 @@
-import { OdfAttributeName } from "../OdfAttributeName";
 import { OdfElementName } from "../OdfElementName";
 
+/**
+ * Utility class for dealing with styles.
+ *
+ * @since 0.4.0
+ */
 export class StyleHelper {
-  public static getStyleElement(document: Document, family: string, name: string): Element {
-    const automaticStylesElement = StyleHelper.getAutomaticStylesElement(document);
-
-    if (automaticStylesElement.childNodes.length > 0) {
-      /* tslint:disable-next-line:prefer-for-of*/
-      for (let i = 0; i < automaticStylesElement.childNodes.length; i++) {
-        const existingStyleElement = automaticStylesElement.childNodes[i] as Element;
-        const nameAttribute = existingStyleElement.attributes.getNamedItem(OdfAttributeName.StyleName);
-        if (nameAttribute !== null && nameAttribute.value === name) {
-          return existingStyleElement;
-        }
-      }
-    }
-
-    const styleElement = document.createElement(OdfElementName.StyleStyle);
-    automaticStylesElement.appendChild(styleElement);
-    styleElement.setAttribute(OdfAttributeName.StyleFamily, family);
-    styleElement.setAttribute(OdfAttributeName.StyleName, name);
-
-    return styleElement;
-  }
-
   /**
    * Returns the `automatic-styles` element of the document.
    * If there is no such element yet, it will be created.
    *
    * @param {Document} document The XML document
    * @returns {Element} The documents `automatic-styles` element
+   * @since 0.4.0
    */
-  private static getAutomaticStylesElement(document: Document): Element {
+  public static getAutomaticStylesElement(document: Document): Element {
     const rootNode = document.firstChild as Element;
 
     const automaticStylesElements = rootNode.getElementsByTagName(OdfElementName.OfficeAutomaticStyles);
@@ -48,6 +31,7 @@ export class StyleHelper {
    *
    * @param {Document} document The XML document
    * @returns {Element} The newly created `automatic-styles` element
+   * @since 0.4.0
    */
   private static createAutomaticStylesElement(document: Document): Element {
     const rootNode = document.firstChild as Element;