+++ src/style/TextProperties.ts
@@ -1,9 +1,7 @@
-import { createHash } from "crypto";
 import { OdfAttributeName } from "../OdfAttributeName";
 import { OdfElementName } from "../OdfElementName";
 import { Color } from "./Color";
-import { ITextStyle } from "./ITextStyle";
-import { StyleHelper } from "./StyleHelper";
+import { ITextProperties } from "./ITextProperties";
 import { Typeface } from "./Typeface";
 
 const MINIMAL_FONT_SIZE = 2;
@@ -13,10 +11,10 @@ const DEFAULT_TYPEFACE = Typeface.Normal;
 /**
  * This class represents the style of some text.
  *
- * @extends ITextStyle
+ * @extends ITextProperties
  * @since 0.4.0
  */
-export class TextStyle implements ITextStyle {
+export class TextProperties implements ITextProperties {
   private color: Color | undefined;
   private fontSize: number;
   private typeface: Typeface;
@@ -74,39 +72,19 @@ export class TextStyle implements ITextStyle {
   }
 
   /**
-   * Returns the name of the style.
-   * The name is computed to make sure equal styles feature equal names and reflects the current configuration.
-   *
-   * @returns {string} The name of the style
-   * @since 0.4.0
-   */
-  public getName(): string {
-    const hash = createHash("md5");
-
-    if (this.color !== undefined) {
-      hash.update(this.color.toHex());
-    }
-    hash.update(this.fontSize.toString());
-    hash.update(this.typeface.toString());
-
-    return hash.digest("hex");
-  }
-
-  /**
    * Transforms the text style into Open Document Format.
    *
    * @param {Document} document The XML document
+   * @param {Element} parent The parent node in the DOM (`style:style`)
    * @since 0.4.0
    */
-  public toXML(document: Document, styleName: string): void {
+  public toXml(document: Document, parent: Element): void {
     if (this.isDefault() === true) {
       return;
     }
 
-    const styleElement = StyleHelper.getStyleElement(document, "paragraph", styleName);
-
     const textPropertiesElement = document.createElement(OdfElementName.StyleTextProperties);
-    styleElement.appendChild(textPropertiesElement);
+    parent.appendChild(textPropertiesElement);
 
     this.setColorAttribute(textPropertiesElement);
     this.setFontSizeAttribute(textPropertiesElement);